Dashboard links: option to add local timezone to URL
