+++ public/app/features/dashboard/components/DashLinks/editor.html
@@ -161,6 +161,13 @@
         ></gf-form-switch>
         <gf-form-switch
           class="gf-form"
+          label="Timezone"
+          checked="ctrl.link.timezone"
+          switch-class="max-width-6"
+          label-class="width-9"
+        ></gf-form-switch>
+        <gf-form-switch
+          class="gf-form"
           label="Open in new tab"
           checked="ctrl.link.targetBlank"
           switch-class="max-width-6"
